test with auto-update as separate module
Moved auto-update functionality to separate module to improve implementation to other tools.
